test: refactor scalar mul gates
